bw/pet: send pet expressions

+++ BinWeevils.Protocol/Str/Pet/PetAction.cs
@@ -7,7 +7,7 @@ namespace BinWeevils.Protocol.Str.Pet
         [StrField] public uint m_petID;
         [StrField] public int m_actionID;
         [StrField] public string m_stateVars;
-        [StrField] public int m_broadcastSwitch;
+        [StrField] public byte m_broadcastSwitch;
         [StrField] public string m_extraParams;
     }
